feat: split json related logic into json package
