+++ tuiexporter/internal/tui/component/modal_test.go
@@ -8,34 +8,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func Test_prettyJSON(t *testing.T) {
-	tests := []struct {
-		name     string
-		input    string
-		expected string
-	}{
-		{
-			name:  "valid JSON object",
-			input: `{"foo":"bar"}`,
-			expected: `{
-  "foo": "bar"
-}`,
-		},
-		{
-			name:     "invalid JSON",
-			input:    `{invalid}`,
-			expected: `{invalid}`,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := prettyJSON(tt.input)
-			assert.Equal(t, tt.expected, result)
-		})
-	}
-}
-
 func Test_attachModalForTreeAttributes(t *testing.T) {
 	tests := []struct {
 		name          string
